extracted the message receiving code from each pcd example page into a hook in the passport-interface package
