+++ apps/consumer-client/pages/zuzalu-examples/group-proof.tsx
@@ -1,8 +1,8 @@
 import {
   requestZuzaluMembershipUrl,
+  useListenToPCDMessage,
   useSemaphorePassportProof,
 } from "@pcd/passport-interface";
-import { useEffect, useState } from "react";
 import { CodeLink, CollapsableCode, HomeLink } from "../../components/Core";
 import { ExampleContainer } from "../../components/ExamplePage";
 import { IS_PROD, PASSPORT_URL } from "../../src/constants";
@@ -13,32 +13,11 @@ const SEMAPHORE_GROUP_URL = IS_PROD
   : "http://localhost:3002/semaphore/1";
 
 export default function Web() {
-  // Raw string-encoded PCD
-  const [pcdStr, setPcdStr] = useState("");
-
-  // Semaphore Group PCD
-  const { proof, group, valid, error } = useSemaphorePassportProof(
+  const pcdStr = useListenToPCDMessage();
+  const { proof, group, valid } = useSemaphorePassportProof(
     SEMAPHORE_GROUP_URL,
     pcdStr
   );
-
-  useEffect(() => {
-    if (error) {
-      console.error("error using semaphore passport proof", error);
-    }
-  }, [error]);
-
-  // Listen for PCDs coming back from the Passport popup
-  useEffect(() => {
-    window.addEventListener("message", receiveMessage, false);
-    function receiveMessage(ev: MessageEvent<any>) {
-      // This next line is important. Extensions including Metamask apparently
-      // send messages to every page. Ignore those.
-      if (!ev.data.encodedPcd) return;
-      console.log("Received message", ev.data);
-      setPcdStr(ev.data.encodedPcd);
-    }
-  }, []);
 
   return (
     <>
