+++ packages/passport-interface/index.ts
@@ -1,6 +1,7 @@
 export * from "./src/EncryptedStorage";
 export * from "./src/Participant";
 export * from "./src/PassportInterface";
+export * from "./src/PCDIntegration";
 export * from "./src/RequestTypes";
 export * from "./src/SemaphoreGroupIntegration";
 export * from "./src/SemaphoreSignatureIntegration";
